+++ Changelog.md
@@ -10,6 +10,7 @@
 ### Krokiet
 - Ability to select multiple items with mouse and keyboard - [#1478](https://github.com/qarmin/czkawka/pull/1478)
 - Added sort buttons - [#1501](https://github.com/qarmin/czkawka/pull/1501)
+- Remembering window size - [TODO]
 
 ### External
 – There is a new, unofficial Tauri-based frontend for Czkawka - [Czkawka Tauri](https://github.com/shixinhuang99/czkawka-tauri)